Add permissions for user routes

+++ user/user.routes.ts
@@ -2,26 +2,37 @@ import express, {Router} from 'express'
 
 import {User} from './user.schema'
 import jwt from "jsonwebtoken"
+import passport from "../middleware/passport"
 
 export const router = Router()
 
-router.get('/', async (req, res) => {
+router.get('/', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const users = await User.find()
-        res.json(users)
+        if (req.user?.role === 'admin') {
+            const users = await User.find()
+            res.json(users)
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
+        }
     } catch (e) {
         res.status(500).json({message: 'something went wrong'})
     }
 })
 
-router.get('/:id', async (req, res) => {
+router.get('/:id', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const user = await User.findById(req.params.id)
-        if (!user) {
-            res.status(404).json({message: 'User does not exist'})
-            return
+        if (req.user?.role === 'admin' || JSON.stringify(req.params.id) !== JSON.stringify(req.user?._id)) {
+            const user = await User.findById(req.params.id)
+            if (!user) {
+                res.status(404).json({message: 'User does not exist'})
+                return
+            }
+            res.status(200).json(user)
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
         }
-        res.status(200).json(user)
     } catch (e) {
         res.status(404).json(e)
     }
@@ -52,34 +63,43 @@ router.post('/',
         }
     })
 
-router.patch('/:id',
+router.patch('/:id', passport.authenticate('jwt', { session: false }),
     async (req: express.Request, res: express.Response) => {
         try {
-            const params = req.body
-            const user = await User.findById(req.params.id)
+            if (req.user?.role === 'admin' || JSON.stringify(req.params.id) !== JSON.stringify(req.user?._id)) {
+                const params = req.body
+                const user = await User.findById(req.params.id)
 
-            if (!user) {
-                res.status(404).json({message: 'User does not exist'})
-                return
+                if (!user) {
+                    res.status(404).json({message: 'User does not exist'})
+                    return
+                }
+
+                Object.assign(user, params)
+                await user.save()
+                res.status(200).json('User updated')
+            } else {
+                res.status(403).json({message: 'You don\'t have permission to access this resource'})
             }
-
-            Object.assign(user, params)
-            await user.save()
-            res.status(200).json('User updated')
         } catch (e) {
             res.status(404).json({message: 'Bad request'})
         }
     })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const user = await User.findById(req.params.id)
-        if (!user) {
-            res.status(404).json({message: 'User does not exist'})
-            return
+        if (req.user?.role === 'admin') {
+            const user = await User.findById(req.params.id)
+            if (!user) {
+                res.status(404).json({message: 'User does not exist'})
+                return
+            }
+            await user.deleteOne()
+            res.status(200).json('User deleted')
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
         }
-        await user.deleteOne()
-        res.status(200).json('User deleted')
     } catch (e) {
         res.status(404).json({message: 'Bad request'})
     }
